fix: Update operation wasn't working

+++ webapp/database/product.py
@@ -1,3 +1,4 @@
+from flask import request
 from . import db
 from typing import List
 from models.product import *
@@ -22,6 +23,10 @@ def get_product(code: str) -> Product | DB_Error:
 
 def create_or_update_product(code: str, name: str) -> Product | DB_Error:
     try:
+        if request.method == "PUT":
+            delete_product(code)
+        else:
+            return {"db_error": f"Product with code={code} already exist"}
         product: Product = Product(code=code, name=name)
         db.session.add(product)
         db.session.commit()
@@ -32,7 +37,9 @@ def create_or_update_product(code: str, name: str) -> Product | DB_Error:
 
 def delete_product(code: str) -> Product | DB_Error:
     try:
-        product: Product = get_product(code)
+        product: Product | DB_Error = get_product(code)
+        if not isinstance(product, Product):
+            return {"db_error": "Product not found"}
         db.session.delete(product)
         db.session.commit()
         return product
